Fix build error: correct type from int[] to float[] for den locations
Den location coordinates are stored as float arrays, not int arrays. This fixes the type mismatch compilation error in LogPlayerLocation.

+++ SysBot.Pokemon/SV/BotRaid/RotatingRaidBotSV.cs
@@ -4227,7 +4227,7 @@ ALwkMx63fBR0pKs+jJ8DcFrcJR50aVv1jfIAQpPIK5G6Dk/4hmV12Hdu5sSGLl40
             .Where(kv => !string.IsNullOrEmpty(kv.Value))
             .Select(kv =>
             {
-                Dictionary<string, int[]>? denLocations = kv.Key switch
+                Dictionary<string, float[]>? denLocations = kv.Key switch
                 {
                     "Blueberry" => blueberryLocations,
                     "Kitakami" => kitakamiLocations,
